Add a test method to get uncompressed data for a model table
